Adding ability for an -option to take variable args after it.
i.e. owping -sender sendaddr [sendserv]

In this example 1 or 2 args. To do this - the arg description lists the max
number of args it will take - and it specifies it as a negative number
in the description record.

This library will eventually have to be documented to some extent - I'm
hoping doxygen is easy...

jeff


git-svn-id: http://anonsvn.internet2.edu/svn/I2util/trunk@6 af504c07-53ac-4b46-a7f1-a0422916b9a2