Tolya - I corrected some problems with the table code, and added some features.
There was a mistake in the init function - you only used one malloc, and
pointed the "buckets" pointer to the memory just past the I2table record.
The problem is that there is nothing making sure that the memory just
past the I2table record is aligned properly for pointers (the type of the
array). There are two ways to fix this - one, add something to the end
of the structure to ensure the next memory is aligned, or just malloc
the buckets array seperately. (the extra malloc seemed easier, so that
is what I did.)

Another mistake in the init function was that you called exit. A library
function should almost never call exit - let the calling environment
handle the errors, just report them. In this case, it is very easy to
just return NULL if there is a problem.

I simplified the cmp function slightly - there was no reason to check for
the MIN of the sizes in the bcmp (which I changed to memcmp) since you
already know the sizes are equal or the first part of the || would have
caused the value to be returned already. (lazy evaluation)

In init - I reordered the primes array check for sizes so you don't have to
have the extra 509, and the bogus INT_MAX in the array. Not really necessary,
but I had a hard time following the logic each time I looked at it before.

hash_close - I made it take an I2table instead of a pointer to an I2table. It
wasn't clear to me why you had it the other way... I also removed the if
checkes before the free'ing of the pointers. (If those pointers were NULL you
would have core dumped from the lines before the if, because those pointers
are dereferenced.)

I added a hash_delete function so you can remove a key/value pair from the
hash. (I added the iterate function earlier - you probably noticed.)

I didn't do it, but I think it would be much cleaner to implement your hash
printing using the hash_iterate function. Then you could take the hash_print
function pointer out of the hash table structure. Also - you could take
the I2binding structure out of the header file since it wouldn't be needed.

Everything else I did was purely cosmetic - defined types for the function
pointers and things along those lines.

jeff


git-svn-id: http://anonsvn.internet2.edu/svn/I2util/trunk@15 af504c07-53ac-4b46-a7f1-a0422916b9a2