Modified the random functions to be relative to a random "context". Modified the OWPErr functions to exclusively use the I2Err functions. (The OWPErr functions are simply a wrapper around the I2Err functions - unfortunately they can't be macros due to the var-args...)
This means if you want to have an application defined error function, you have
to define the I2ErrHandler yourself, and assign it to eh in the OWPContext.

This actually simplified things quite a bit. I would almost like to go
through and remove all the OWPError and OWPErrLine calls and replace them
with I2 varients - may do that later, doesn't seem worth it right now. It
would make the __FILE__ and __LINE__ macro's cleaner.

jeff


git-svn-id: http://anonsvn.internet2.edu/svn/I2util/trunk@29 af504c07-53ac-4b46-a7f1-a0422916b9a2